type up @covector/apply
